[CPL:INPUT] Update Input's existing localized information. (#861)
Update the name of the Romanian Keyboard layout in localized
resources by renaming the existing one as legacy and adding
the new one as standard.

Use the same resource string ID as the one used in input.cpl from
later versions of Windows that do have kbdrost.
This improves the Windows binary compatibility of kbrost's setup.

CORE-15056