[MPR] Drop a nasty ReactOS specific hack
That hack (cf. previous commit) was used to restore all
the saved connections. But, because it was implemented
in the wrong place, each time the MPR DLL was attached,
it was trying to restore connections.

With that commit and the previous one, now, ReactOS has
a correct behavior (it tries to bring back connections
only at session opening) and will reduce "workload" for
applications attaching MPR dll.

CORE-15310