[SHELL32] Improve CORE-15295 IDD_FILE_PROPERTIES
addendum to 0.4.11-dev-673-g
04e1263921c0a4ee9b5ec10bb02884146fcca47a

Actually this control must not statically consume width of 170.
There are two different versions of that dialog.
One has a "change"-button. In this case the edit must be only 100 wide.
In case we have the version without the change button, we resize it to have
170 width (same as its neighbor).

This commit is shared work of
Stanislav Motylkov, Giannis Adamopoulos and Joachim Henze.

Commit supersedes PR#1056.