[PORTCLS] Zero memory in operator new. CORE-16157
Fixes a regression from 99fa38809fd8ed38d4d59fa1fcf3babb3216b895, which
replaced the kcom.h/stdunk.h versions (which zero memory) with local
implementations (which don't).

Standard C++ does not guarantee that memory is zeroed but several classes
rely on this, in particular for their m_Ref and m_bInitialized members.

Ideally the constructors should be fixed to initialize all required members,
but that task is error-prone so for now we simply restore the previous
behavior.