[INCLUDE:HOST] Revert a bug that has been sneaked in by Arch Blackmann in 9e9ba41f (r36910) :P
The "TypeOffset" thing was just an informative comment to tell that the
data that follows after the IMAGE_BASE_RELOCATION header is an arbitrary
array of WORDs describing packed (Type + Offset)s.

Having the header structure containing that spurious "TypeOffset" was
breaking all the code that was basing on expected size of the
IMAGE_BASE_RELOCATION structure in order to apply relocations (typically
this would mean that the first 2 relocations described by it would not
be applied).

Hopefully this bug only hit the host-tools, and not the OS itself :)